fix(css): separate root vs child theme class names for Lynx dark mode
- Use `seed-user-color-scheme-light/dark` for root-level resolved theme
  (set by rsbuild plugin based on system theme or color mode config)
- Keep `seed-color-mode-light-only/dark-only` for child-level theme overrides
- Update postcss-lynx-compat selectorMappings for 1:1 attribute-to-class conversion
- Update getThemeClassName() to return `seed-user-color-scheme-*`
- Add text slot skip logic for `seed-user-color-scheme-*` selectors
- Split postcss-lynx-compat tests into 6 focused files
- Add Lynx theming documentation
- Regenerate all `.lynx.css` files
