+++ packages/css/base.lynx.css
@@ -381,7 +381,7 @@
   --seed-timing-function-exit-expressive: cubic-bezier(.35, 0, .95, .55);
 }
 
-:root, :root.seed-color-mode-light-only, .seed-color-mode-light-only {
+:root, :root.seed-user-color-scheme-light, :root.seed-color-mode-light-only, .seed-color-mode-light-only {
   --seed-color-palette-gray-00: #fff;
   --seed-color-palette-gray-100: #f7f8f9;
   --seed-color-palette-gray-200: #f3f4f5;
@@ -592,7 +592,7 @@
   --seed-shadow-s3: 0px 4px 16px 0px #0000001f;
 }
 
-:root.seed-color-mode-dark-only, .seed-color-mode-dark-only {
+:root.seed-user-color-scheme-dark, :root.seed-color-mode-dark-only, .seed-color-mode-dark-only {
   --seed-color-palette-gray-00: #000;
   --seed-color-palette-gray-100: #16171b;
   --seed-color-palette-gray-200: #1d2025;
