+++ packages/rsbuild-plugin/src/lynx-theming.ts
@@ -3,15 +3,18 @@ import type { ColorMode } from "@seed-design/css/theming";
 /**
  * Determine the theme class name for a Lynx `<page>` element.
  *
+ * Uses `seed-user-color-scheme-*` to represent the resolved theme on root,
+ * while `seed-color-mode-*-only` is reserved for child-level theme overrides.
+ *
  * @param colorMode - Plugin color mode option ("system" | "light-only" | "dark-only")
  * @param systemTheme - Value of `lynx.__globalProps.theme` ("Light" | "Dark")
- * @returns "seed-color-mode-light-only" or "seed-color-mode-dark-only"
+ * @returns "seed-user-color-scheme-light" or "seed-user-color-scheme-dark"
  */
 export function getThemeClassName(colorMode: ColorMode, systemTheme: string | undefined): string {
-  if (colorMode === "light-only") return "seed-color-mode-light-only";
-  if (colorMode === "dark-only") return "seed-color-mode-dark-only";
+  if (colorMode === "light-only") return "seed-user-color-scheme-light";
+  if (colorMode === "dark-only") return "seed-user-color-scheme-dark";
   // "system" → follow device dark mode setting
   return systemTheme?.toLowerCase() === "dark"
-    ? "seed-color-mode-dark-only"
-    : "seed-color-mode-light-only";
+    ? "seed-user-color-scheme-dark"
+    : "seed-user-color-scheme-light";
 }
